+++ CHANGELOG.md
@@ -40,7 +40,7 @@ v4.6
 - Methods are now available in `OpenSim::Component` for generating a list of all `ModelingOption`s and `DiscreteVariable`s in an `OpenSim::Model` or other `Component`. (#3745)
 - `OpenSim::Component` accessor methods for `ModelingOption`s and `DiscreteVariables`s now accept both absolute and relative component paths. (#3745)
 - `DiscreteVariables` in OpenSim can now be a range of numerical types, including `bool`, `int`, `double`, `Vec2`, `Vec3`, ..., `Vec6`, and `Quaternion`. (#3745)
-- `DiscreteVariable`s and `ModelingOption`s allocated natively in Simbody can now be added to an `OpenSim::Component` and accessed via its `Component` API. (#3745)
+- `DiscreteVariable`s and `ModelingOption`s allocated natively in Simbody can now be added to an `OpenSim::Component` and accessed via its `Component` API. To support this capability, `getDiscreteVariableIndex()` has been replaced by `getDiscreteVariableIndexes()` which returns both the index of the discrete variable and the index of the `SimTK::Subsystem` to which the descrete variable belongs. (#3745)
 - Computationally efficient methods are now available for extracting the time histories of individual state variables, discrete states, and modeling options from a state trajectory (i.e., a `SimTK::Array_<SimTK::State>`). Collectively, these methods form the basis for performing a comprehensive serialzation of a state trajectory to file. (#3745)
 - Computationally efficient methods are now available for building a state trajectory (i.e., a `SimTK::Array_<SimTK::State>`) from the time histories of individual state variables, discrete states, and modeling options. Collectively, these methods form the basis for performing a comprehenvise deserialization of a states trajectory from file. (#3745)